Remove Q_SCRIPTABLE macro, otherwise it no workie.
svn path=/trunk/KDE/kdeaccessibility/kttsd/; revision=552420

+++ kttsd/speechdata.cpp
@@ -918,7 +918,7 @@ int SpeechData::getTextJobState(const uint jobNum)
     mlJob* job = findJobByJobNum(jobNum);
     int temp;
     if (job)
-        temp = job->state;
+        temp = (int)job->state;
     else
         temp = -1;
     return temp;
